requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
contourpy==1.3.3
cuda-bindings==13.2.0
cuda-pathfinder==1.5.3
cuda-toolkit==13.0.2
cycler==0.12.1
distro==1.9.0
fastapi==0.136.1
filelock==3.29.0
fonttools==4.62.1
fsspec==2026.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
Jinja2==3.1.6
jiter==0.14.0
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.9
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
openai==2.32.0
opencv-python==4.13.0.92
packaging==26.2
pillow==12.2.0
polars==1.40.1
polars-runtime-32==1.40.1
psutil==7.2.2
pydantic==2.13.3
pydantic_core==2.46.3
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.26
PyYAML==6.0.3
requests==2.33.1
scipy==1.17.1
setuptools==81.0.0
six==1.17.0
sniffio==1.3.1
starlette==1.0.0
sympy==1.14.0
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
triton==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.41
ultralytics-thop==2.0.19
urllib3==2.6.3
uvicorn==0.46.0
